@@ -6,7 +6,7 @@
   "main": "dist/main.js",
   "type": "module",
   "volta": {
-    "node": "20.7.0",
+    "node": "20.8.0",
     "pnpm": "8.8.0"
   },
   "scripts": {
@@ -26,7 +26,7 @@
   },
   "devDependencies": {
     "@rollup/plugin-typescript": "^11.1.4",
-    "@types/node": "^20.7.1",
+    "@types/node": "^20.7.2",
     "autocannon": "^7.12.0",
     "rollup": "^3.29.4",
     "rollup-plugin-delete": "^2.0.0",
