@@ -1,11 +1,13 @@
-'use strict'
-const { ThreadWorker, KillBehaviors } = require('../../../lib')
+import { KillBehaviors, ThreadWorker } from '../../../lib/index.js'
 
+/**
+ *
+ */
 function error () {
   throw new Error('Error Message from ThreadWorker')
 }
 
-module.exports = new ThreadWorker(error, {
+export default new ThreadWorker(error, {
   killBehavior: KillBehaviors.HARD,
   maxInactiveTime: 500
 })
