@@ -1,6 +1,8 @@
-export const TaskFunctions = {
+const TaskFunctions = {
   jsonIntegerSerialization: 'jsonIntegerSerialization',
   fibonacci: 'fibonacci',
   factorial: 'factorial',
   readWriteFiles: 'readWriteFiles'
 }
+
+module.exports = { TaskFunctions }
