@@ -3,11 +3,11 @@ import {
   WorkerTypes,
   availableParallelism
 } from '../../lib/index.mjs'
-import { TaskFunctions } from '../benchmarks-types.mjs'
+import { TaskFunctions } from '../benchmarks-types.js'
 import {
   buildPoolifierPool,
   runPoolifierPoolBenchmark
-} from '../benchmarks-utils.mjs'
+} from '../benchmarks-utils.js'
 
 const poolSize = availableParallelism()
 const taskExecutions = 1
@@ -30,6 +30,26 @@ await runPoolifierPoolBenchmark(
 await runPoolifierPoolBenchmark(
   'Poolifier DynamicThreadPool',
   buildPoolifierPool(WorkerTypes.thread, PoolTypes.dynamic, poolSize),
+  {
+    taskExecutions,
+    workerData
+  }
+)
+
+// FixedClusterPool
+await runPoolifierPoolBenchmark(
+  'Poolifier FixedClusterPool',
+  buildPoolifierPool(WorkerTypes.cluster, PoolTypes.fixed, poolSize),
+  {
+    taskExecutions,
+    workerData
+  }
+)
+
+// DynamicClusterPool
+await runPoolifierPoolBenchmark(
+  'Poolifier DynamicClusterPool',
+  buildPoolifierPool(WorkerTypes.cluster, PoolTypes.dynamic, poolSize),
   {
     taskExecutions,
     workerData
